backend/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 圖片處理
pillow==10.1.0

# AI 服務
google-generativeai==0.3.1

# 資料庫
supabase==2.0.3

# Google Sheets 整合
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# WebSocket
websockets==12.0

# 資料驗證
pydantic==2.5.0
pydantic-settings==2.1.0

# 環境變數
python-dotenv==1.0.0

# HTTP 客戶端
httpx==0.25.2

# 工具
python-jose[cryptography]==3.3.0  # JWT 處理
passlib[bcrypt]==1.7.4  # 密碼處理
